♻️ 资产类型 BUILTIN_OPTIONS 元数据折进 AssetTypeDefinition,图标统一品牌图 #130
