+++ frontend/src/lib/assetTypes/k8s.ts
@@ -1,10 +1,13 @@
-import { Container } from "lucide-react";
+import { KubernetesIcon } from "@/components/asset/brand-icons";
 import { registerAssetType } from "./_register";
 import { K8sDetailInfoCard } from "@/components/asset/detail/K8sDetailInfoCard";
 
 registerAssetType({
   type: "k8s",
-  icon: Container,
+  icon: KubernetesIcon,
+  aliases: ["k8s", "kubernetes"],
+  label: "nav.k8s",
+  category: "middleware",
   canConnect: true,
   canConnectInNewTab: false,
   connectAction: "terminal",
