+++ frontend/src/lib/assetTypes/ssh.ts
@@ -1,10 +1,13 @@
-import { Server } from "lucide-react";
+import { Monitor } from "lucide-react";
 import { registerAssetType } from "./_register";
 import { SSHDetailInfoCard } from "@/components/asset/detail/SSHDetailInfoCard";
 
 registerAssetType({
   type: "ssh",
-  icon: Server,
+  icon: Monitor,
+  aliases: ["ssh"],
+  label: "nav.ssh",
+  category: "servers",
   canConnect: true,
   canConnectInNewTab: true,
   connectAction: "terminal",
