+++ frontend/src/lib/assetTypes/types.ts
@@ -21,9 +21,18 @@ export interface PolicyDefinition {
   fields: PolicyFieldDef[];
 }
 
+/** 语义分组（资产类型选择器展示用）。 */
+export type AssetTypeCategory = "servers" | "databases" | "middleware" | "extension";
+
 export interface AssetTypeDefinition {
   type: string;
   icon: ComponentType<{ className?: string; style?: React.CSSProperties }>;
+  /** 所有应匹配此类型的 `asset.Type` 值（含历史别名）。 */
+  aliases: string[];
+  /** 选择器展示标签的 i18n key（默认命名空间），如 `nav.ssh`。 */
+  label: string;
+  /** 选择器语义分组。 */
+  category: AssetTypeCategory;
   canConnect: boolean;
   canConnectInNewTab: boolean;
   connectAction: "terminal" | "query";
